requirements.txt
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
Pillow==8.0.1
pymongo==3.11.1
python-pptx==0.6.18
Werkzeug==1.0.1
XlsxWriter==1.3.7
